Smooth scroll the visual viewport on the compositor.
Originally, only the layoutViewport was run on the compositor for smooth scroll
animation. This CL simultaneously runs the smooth scroll animation for both the
visual and layout viewport on the compositor. Initially, the idea was to run the
visual viewport animation with a delay, but running them together looks visually
okay and keeps the code simple.

Note that this CL should only affect unshipped code (CSSOM Smooth Scroll,
which is behind a runtime enabled flag).

BUG=492924

Review URL: https://codereview.chromium.org/1340983003

git-svn-id: svn://svn.chromium.org/blink/trunk@202384 bbb929c8-8fbe-4397-9dbb-9b2b20218538