Oilpan: like Vector<T>, restrict Deque<T> to certain part objects.
Impose the limitations put on Vector<T> on Deque<T> also. They're
both backed by a vector buffer where unused slots are zero'ed so as
to be safe when (conservatively) tracing the buffer contents or when
finalizing it.

In order to use a designated zero for cleared elements, certain
conditions have to be placed on the kinds of T we can allow when
T is a part object annotated with ALLOW_ONLY_INLINE_ALLOCATION().
Impose these statically by way of static_assert()s.

R=tkent,haraken
BUG=420515

Review URL: https://codereview.chromium.org/1051853003

git-svn-id: svn://svn.chromium.org/blink/trunk@202574 bbb929c8-8fbe-4397-9dbb-9b2b20218538